Add ability to skip verifying TLS certs
This commit modifies the requests package to add the ability to not
verify TLS certificates when making HTTP requests to HTTPS servers.

Enabling the option is done by calling the NoTLSVerify() method on the
requests client object, providing a boolean value to enable or disable
the feature. Modifying this option can be performed at runtime and
doesn't have to happen when first creating the client.