Introduce requests - an HTTP client
This commit introduces a new utility called "requests", which is a
client for HTTP requests. This client features a builder-style method of
issuing requests, and includes various safe-guards meant to prevent
leaking connections and making other mistakes that are common when
manually writing requests with net/http.

The package is tailored towards working with web services and RESTful
APIs but is meant to work well with any website or net-facing
application.

To use, one must create an HTTPClient object (through the NewClient()
function) for a certain web service. The user provides a base URL which
will be used for all requests, and other options common for all requests
(e.g. authentication, expected content type, etc.). Users then create
HTTP requests through the client object (through the NewRequest()
method). The package ensures that request bodies are properly read in
full, closed, and copied into a variable (optionally decoded from JSON
or other types).