Bugfix: retries have no body; header timeout short
When the library retries requests, subsequent requests will have no
body. This is because the body is provided as an io.Reader, which is
read to completion upon the first attempt, and then subsequent attempts
can't read anything. To fix, every attempt will receive a new reader.

Also included is a removal of the response header timeout, as this is
only causing issues (the default Go transport also does not have this
timeout).