Bugfix: slash added when making request to an empty path
When the base URL provided to the client is already a unique resource
locator (URL), doing a request to an empty path may result in a 404 Not
Found, because the client automatically added a slash.

This commit modifies the behavior to only add a slash prefix if the path
provided to NewRequest is not empty.