Add support for subscribing to Server-Sent Events
This commit introduces support for subscribing to Server-Sent Events.
To do so, you create a request as normal, but use `Subscribe` instead
of `Run` (or `RunContext`). The `Subscribe` method requires a context
object, and a `[]byte` channel.

The library will issue the request, then start reading messages from the
server in a goroutine. Every message will be sent to the provided
channel as-is. When all messages have been read, the channel is closed.