Add ability to pre-process a request's body
Users can now provide a body processor function that can
pre-process the body before it is sent, potentially modifying it.
The processor function is called right before compressing the
body (if compression is enabled) and issuing the request, so when
using a JSON body, for example, the function is called after the
body has been serialized to JSON.

The project now requires Go version 1.16 and above.